backend/requirements.txt
fastapi
uvicorn[standard]
typing-extensions
flask
flask-cors
python-dotenv
langgraph
pydantic
langchain-google-genai
langchain-core